chore(tests/storage): moved tests to match src
