+++ test/types/Storage/methods/addUTXOToAddress.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const addUTXOToAddress = require('../../../src/types/Storage/methods/addUTXOToAddress');
+const addUTXOToAddress = require('../../../../src/types/Storage/methods/addUTXOToAddress');
 
 describe('Storage - addUTXOToAddress', () => {
   it('should add a new UTXO to an address', () => {
