+++ test/types/Storage/methods/announce.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const announce = require('../../../src/types/Storage/methods/announce');
+const announce = require('../../../../src/types/Storage/methods/announce');
 
 describe('Storage - announce', () => {
   // TODO
