+++ test/types/Storage/methods/importAddresses.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const importAddresses = require('../../../src/types/Storage/methods/importAddresses');
+const importAddresses = require('../../../../src/types/Storage/methods/importAddresses');
 
 describe('Storage - importAddresses', () => {
   it('should import an array of addresses', () => {
