+++ test/types/Storage/methods/importSingleAddress.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const importSingleAddress = require('../../../src/types/Storage/methods/importSingleAddress');
+const importSingleAddress = require('../../../../src/types/Storage/methods/importSingleAddress');
 
 describe('Storage - importSingleAddress', () => {
   it('should import a SingleAddress (non-bip44)', () => {
