+++ test/types/Storage/methods/rehydrateState.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const rehydrateState = require('../../../src/types/Storage/methods/rehydrateState');
+const rehydrateState = require('../../../../src/types/Storage/methods/rehydrateState');
 
 describe('Storage - rehydrateState', () => {
   it('should rehydrate the state', () => {
