+++ test/types/Storage/methods/saveState.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const saveState = require('../../../src/types/Storage/methods/saveState');
+const saveState = require('../../../../src/types/Storage/methods/saveState');
 
 describe('Storage - saveState', () => {
   it('should state the state', () => {
