+++ test/types/Storage/methods/searchAddress.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const searchAddress = require('../../../src/types/Storage/methods/searchAddress');
+const searchAddress = require('../../../../src/types/Storage/methods/searchAddress');
 
 describe('Storage - searchAddress', () => {
   it('should search an address', () => {
