+++ test/types/Storage/methods/searchAddressWithTx.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const searchAddressesWithTx = require('../../../src/types/Storage/methods/searchAddressesWithTx');
+const searchAddressesWithTx = require('../../../../src/types/Storage/methods/searchAddressesWithTx');
 
 describe('Storage - searchAddressesWithTx', () => {
   it('should search an address having a specific tx', () => {
