+++ test/types/Storage/methods/searchTransaction.js
@@ -1,7 +1,7 @@
 const { expect } = require('chai');
 const Dashcore = require('@dashevo/dashcore-lib');
-const searchTransaction = require('../../../src/types/Storage/methods/searchTransaction');
-const transactionsFixtures = require('../../fixtures/transactions');
+const searchTransaction = require('../../../../src/types/Storage/methods/searchTransaction');
+const transactionsFixtures = require('../../../fixtures/transactions');
 
 const { faa430b0fe84a074d981e6fa3995a13363478415ca029a12f6432bf3d90dfa60, fd7c727155ef67fd5c1d54b73dea869e9690c439570063d6e96fec1d3bba450e } = transactionsFixtures.valid.mainnet;
 describe('Storage - searchTransaction', () => {
