+++ test/types/Storage/methods/updateTransaction.js
@@ -1,6 +1,6 @@
 const { expect } = require('chai');
-const updateTransaction = require('../../../src/types/Storage/methods/updateTransaction');
-const orangeWStore = require('../../fixtures/walletStore').valid.orange.store;
+const updateTransaction = require('../../../../src/types/Storage/methods/updateTransaction');
+const orangeWStore = require('../../../fixtures/walletStore').valid.orange.store;
 
 describe('Storage - updateTransaction', () => {
   it('should throw on failed update', () => {
